+++ src/types/rx-error.d.ts
@@ -30,10 +30,11 @@ export interface RxErrorParameters {
     readonly childpath?: string;
     readonly obj?: any;
     readonly document?: any;
-    readonly schema?: RxJsonSchema | RxSchema;
+    readonly schema?: Readonly<RxJsonSchema | RxSchema>;
     readonly schemaObj?: any;
     readonly pluginKey?: string;
-    readonly finalDoc?: any;
+    readonly originalDoc?: Readonly<any>;
+    readonly finalDoc?: Readonly<any>;
     readonly regex?: string;
     readonly fieldName?: string;
     readonly id?: string;